Getting Started, Adding a Blog to your site
Getting Started -
1. Added a section explaining how to change the zones that are visible
on a page.  The reason for making such a big change was a. It's a common
task and b. When the user adds a new page they will be asked if they
want to add a new layer for that page.  The concept of layers and what
they are used for should be introduced before the user is asked to make
this decision.

2. Removed the section about adding a blog since this subject matter is
covered in detail in it's own topic.

3. Updated all of the screenshots to Orchard 1.6

Adding a blog to your website -
1. Broke the topic into multiple sections with headers to make it easier
to navigate
2. Updated screenshots to Orchard 1.6

Index -
1. Removed the topic "Adding a Page" as this same material is covered in
detail in the "Getting Started" topic